requirements.txt
# Screen capture & image processing
mss>=9.0
opencv-python>=4.9
numpy>=1.26
Pillow>=10.0
imagehash>=4.3

# OCR
paddlepaddle-gpu>=2.6
paddleocr>=2.8

# Windows API
pywin32>=306

# LLM API
openai>=1.50

# UI
PyQt6>=6.7

# Data
httpx>=0.27
rapidfuzz>=3.6
pydantic>=2.7
pydantic-settings>=2.3

# Config
pyyaml>=6.0
python-dotenv>=1.0

# Logging
loguru>=0.7

# Dev
ruff
